+++ applications/fishing-map/src/features/search/Search.tsx
@@ -285,7 +285,6 @@ function Search() {
             <label className={styles.title}>{t('search.title', 'Search')}</label>
             <Choice
               options={searchOptions}
-              disabledTooltip={t('search.advancedDisabled')}
               activeOption={activeSearchOption}
               onOptionClick={onSearchOptionChange}
               size="small"
@@ -444,7 +443,7 @@ function Search() {
                               )}
                               {firstTransmissionDate && lastTransmissionDate && (
                                 <div className={cx(styles.property, styles.fullWidth)}>
-                                  <label>{t('vessel.transmission_plural', 'Transmissions')}</label>
+                                  <label>{t('vessel.transmission_other', 'Transmissions')}</label>
                                   <span>
                                     from <I18nDate date={firstTransmissionDate} /> to{' '}
                                     <I18nDate date={lastTransmissionDate} />