+++ applications/fishing-map/src/features/workspace/workspace.slice.ts
@@ -138,7 +138,10 @@ export const fetchWorkspaceThunk = createAsyncThunk(
 
 export const saveCurrentWorkspaceThunk = createAsyncThunk(
   'workspace/saveCurrent',
-  async (defaultName: string, { dispatch, getState }) => {
+  async (
+    { name: defaultName, createAsPublic }: { name: string; createAsPublic: boolean },
+    { dispatch, getState }
+  ) => {
     const state = getState() as RootState
     const mergedWorkspace = selectCustomWorkspace(state)
 
@@ -155,8 +158,7 @@ export const saveCurrentWorkspaceThunk = createAsyncThunk(
               body: {
                 ...mergedWorkspace,
                 name,
-                // TODO make this optional for admins
-                public: false,
+                public: createAsPublic,
               },
             } as FetchOptions<WorkspaceUpsert<WorkspaceState>>
           )
@@ -191,13 +193,12 @@ export const saveCurrentWorkspaceThunk = createAsyncThunk(
 
 export const updatedCurrentWorkspaceThunk = createAsyncThunk(
   'workspace/updatedCurrent',
-  async (workspaceId: string, { dispatch, getState }) => {
+  async (workspace: WorkspaceUpsert<WorkspaceState>, { dispatch, getState }) => {
     const state = getState() as RootState
     const version = selectVersion(state)
-    const workspace = selectCustomWorkspace(state)
 
     const workspaceUpdated = await GFWAPI.fetch<Workspace<WorkspaceState>>(
-      `/${version}/workspaces/${workspaceId}`,
+      `/${version}/workspaces/${workspace.id}`,
       {
         method: 'PATCH',
         body: workspace,