+++ applications/vessel-history/src/features/home/Home.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment, useCallback, useEffect, useRef } from 'react'
+import React, { Fragment, useCallback, useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useTranslation } from 'react-i18next'
 import { event as uaEvent } from 'react-ga'
@@ -13,12 +13,7 @@ import VesselListItem from 'features/vessel-list-item/VesselListItem'
 import { useOfflineVesselsAPI } from 'features/vessels/offline-vessels.hook'
 import { selectAll as selectAllOfflineVessels } from 'features/vessels/offline-vessels.slice'
 import SearchPlaceholder, { SearchNoResultsState } from 'features/search/SearchPlaceholders'
-import {
-  selectAdvancedSearchFields,
-  selectHasSearch,
-  selectUrlQuery,
-} from 'routes/routes.selectors'
-import { fetchVesselSearchThunk } from 'features/search/search.thunk'
+import { selectHasSearch } from 'routes/routes.selectors'
 import {
   selectSearchOffset,
   selectSearchResults,
@@ -28,7 +23,7 @@ import {
 import AdvancedSearch from 'features/search/AdvancedSearch'
 import { useUser } from 'features/user/user.hooks'
 import { PROFILE } from 'routes/routes'
-import { useSearchConnect } from 'features/search/search.hooks'
+import { useSearchConnect, useSearchResultsConnect } from 'features/search/search.hooks'
 import { formatVesselProfileId } from 'features/vessels/vessels.utils'
 import styles from './Home.module.css'
 import LanguageToggle from './LanguageToggle'
@@ -45,18 +40,15 @@ const Home: React.FC<LoaderProps> = (): React.ReactElement => {
   const { t } = useTranslation()
   const dispatch = useDispatch()
   const { logout } = useUser()
-  const { onVesselClick, selectedVessels, setSelectedVessels } = useSearchConnect()
+  const { onVesselClick, selectedVessels, setSelectedVessels } = useSearchResultsConnect()
+  const { fetchResults } = useSearchConnect({ onNewSearch: () => setSelectedVessels([]) })
   const searching = useSelector(selectSearching)
-  const query = useSelector(selectUrlQuery)
-  const advancedSearch = useSelector(selectAdvancedSearchFields)
   const hasSearch = useSelector(selectHasSearch)
   const vessels = useSelector(selectSearchResults)
   const offset = useSelector(selectSearchOffset)
   const totalResults = useSelector(selectSearchTotalResults)
   const offlineVessels = useSelector(selectAllOfflineVessels)
   const { dispatchFetchOfflineVessels, dispatchDeleteOfflineVessel } = useOfflineVesselsAPI()
-
-  const promiseRef = useRef<any>()
 
   useEffect(() => {
     dispatchFetchOfflineVessels()
@@ -101,51 +93,31 @@ const Home: React.FC<LoaderProps> = (): React.ReactElement => {
     if (selectedVessel) openVesselProfile(selectedVessel, akaVessels)
   }, [openVesselProfile, selectedVessels, vessels])
 
-  const fetchResults = useCallback(
-    (offset = 0) => {
-
-      if (promiseRef.current) {
-        promiseRef.current.abort()
-      }
-      if (offset === 0) {
-        setSelectedVessels([])
-      }
-      // To ensure the pending action isn't overwritted by the abort above
-      // and we miss the loading intermediate state
-      setTimeout(() => {
-        promiseRef.current = dispatch(
-          fetchVesselSearchThunk({
-            query,
-            offset,
-            ...((advancedSearch ? { advancedSearch } : {}) as any),
-          })
-        )
-      }, 100)
-    },
-    [setSelectedVessels, dispatch, query, advancedSearch]
-  )
   const trackOpenSettings = useCallback(() => {
     uaEvent({
       category: 'Highlight Events',
       action: 'Start highlight events configurations',
       label: JSON.stringify({
-        page: 'home'
-      })
+        page: 'home',
+      }),
     })
   }, [])
 
-  const trackRemoveOffline = useCallback((offlineVessel) => {
-    const now = DateTime.now()
-    const savedOn = DateTime.fromISO(offlineVessel.savedOn);
-    const i = Interval.fromDateTimes(savedOn, now);
-    uaEvent({
-      category: 'Offline Access',
-      action: 'Remove saved vessel for offline view',
-      label: JSON.stringify({ page: 'home' }),
-      value: Math.floor(i.length('days'))
-    })
-    dispatchDeleteOfflineVessel(offlineVessel)
-  }, [dispatchDeleteOfflineVessel])
+  const trackRemoveOffline = useCallback(
+    (offlineVessel) => {
+      const now = DateTime.now()
+      const savedOn = DateTime.fromISO(offlineVessel.savedOn)
+      const i = Interval.fromDateTimes(savedOn, now)
+      uaEvent({
+        category: 'Offline Access',
+        action: 'Remove saved vessel for offline view',
+        label: JSON.stringify({ page: 'home' }),
+        value: Math.floor(i.length('days')),
+      })
+      dispatchDeleteOfflineVessel(offlineVessel)
+    },
+    [dispatchDeleteOfflineVessel]
+  )
 
   useEffect(() => {
     setSelectedVessels([])
@@ -211,12 +183,12 @@ const Home: React.FC<LoaderProps> = (): React.ReactElement => {
                   {selectedVessels.length > 0 && (
                     <div className={styles.bottomActions}>
                       {selectedVessels.length === 1 && (
-                        <Button className={styles.seeVesselBtn} onClick={onSeeVesselClick}>
+                        <Button className={styles.bottomActionsBtn} onClick={onSeeVesselClick}>
                           {t('search.seeVessel', 'See Vessel')}
                         </Button>
                       )}
                       {selectedVessels.length > 1 && (
-                        <Button className={styles.mergeVesselBtn} onClick={onMergeVesselClick}>
+                        <Button className={styles.bottomActionsBtn} onClick={onMergeVesselClick}>
                           {t('search.mergeSelectedVessels', 'Merge Selected Vessels')}
                         </Button>
                       )}