+++ applications/vessel-history/src/features/i18n/react-i18next.d.ts
@@ -1,8 +1,8 @@
 // import the original type declarations
 import 'react-i18next'
 // import all namespaces (for the default language, only)
-import translations from '../../../public/locales/source/translations.json'
-import flags from '../../../public/locales/source/flags.json'
+import type flags from '@globalfishingwatch/i18n-labels/source/flags.json'
+import type translations from '../../../public/locales/source/translations.json'
 
 declare module 'react-i18next' {
   // and extend them!