+++ applications/vessel-history/src/features/profile/components/Info.tsx
@@ -9,10 +9,10 @@ import { Authorization } from '@globalfishingwatch/api-types'
 import { Button, IconButton } from '@globalfishingwatch/ui-components'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { VesselWithHistory } from 'types'
-import I18nDate from 'features/i18n/i18nDate'
+import I18nDate, { I18nSpecialDate } from 'features/i18n/i18nDate'
 import { selectCurrentOfflineVessel } from 'features/vessels/offline-vessels.selectors'
 import { useOfflineVesselsAPI } from 'features/vessels/offline-vessels.hook'
-import { OfflineVessel } from 'types/vessel'
+import { OfflineVessel, VesselFieldLabel } from 'types/vessel'
 import {
   selectDataset,
   selectTmtId,
@@ -21,7 +21,7 @@ import {
   selectVesselProfileId,
 } from 'routes/routes.selectors'
 import { selectEventsForTracks } from 'features/vessels/activity/vessels-activity.selectors'
-import InfoField, { VesselFieldLabel } from './InfoField'
+import InfoField from './InfoField'
 import styles from './Info.module.css'
 import 'react-image-gallery/styles/css/image-gallery.css'
 import Highlights from './Highlights'
@@ -58,13 +58,13 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
 
   const onDeleteClick = async (data: OfflineVessel) => {
     const now = DateTime.now()
-    const savedOn = DateTime.fromISO(data.savedOn);
-    const i = Interval.fromDateTimes(savedOn, now);
+    const savedOn = DateTime.fromISO(data.savedOn)
+    const i = Interval.fromDateTimes(savedOn, now)
     uaEvent({
       category: 'Offline Access',
       action: 'Remove saved vessel for offline view',
       label: JSON.stringify({ page: 'vessel detail' }),
-      value: Math.floor(i.length('days'))
+      value: Math.floor(i.length('days')),
     })
     setLoading(true)
     await dispatchDeleteOfflineVessel(data)
@@ -78,8 +78,8 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
       action: 'Save vessel for offline view',
       label: JSON.stringify({
         gfw: vesselId,
-        tmt: vesselTmtId
-      })
+        tmt: vesselTmtId,
+      }),
     })
     await dispatchCreateOfflineVessel({
       vessel: {
@@ -109,7 +109,7 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
     <Fragment>
       <div className={styles.infoContainer}>
         {vessel && (
-          <Fragment>
+          <div className={styles.imageAndFields}>
             {imageList.length > 0 && (
               <ImageGallery
                 items={imageList}
@@ -193,9 +193,17 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
                     {auth.source}{' '}
                     <Fragment>
                       {t('common.from', 'from')}{' '}
-                      {auth.startDate ? <I18nDate date={auth.startDate} /> : DEFAULT_EMPTY_VALUE}{' '}
+                      {auth.startDate ?? auth.originalStartDate ? (
+                        <I18nSpecialDate date={auth.startDate ?? auth.originalStartDate} />
+                      ) : (
+                        DEFAULT_EMPTY_VALUE
+                      )}{' '}
                       {t('common.to', 'to')}{' '}
-                      {auth.endDate ? <I18nDate date={auth.endDate} /> : DEFAULT_EMPTY_VALUE}
+                      {auth.endDate ?? auth.originalEndDate ? (
+                        <I18nSpecialDate date={auth.endDate ?? auth.originalEndDate} />
+                      ) : (
+                        DEFAULT_EMPTY_VALUE
+                      )}
                     </Fragment>
                   </p>
                 ))}
@@ -259,7 +267,7 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
                 valuesHistory={[]}
               ></InfoField>
             </div>
-          </Fragment>
+          </div>
         )}
         <div className={styles.actions}>
           {vessel && offlineVessel && (