+++ applications/vessel-history/src/features/profile/components/InfoField.tsx
@@ -4,32 +4,10 @@ import { useTranslation } from 'react-i18next'
 import I18nDate from 'features/i18n/i18nDate'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { ValueItem } from 'types'
-import styles from './Info.module.css'
+import { VesselFieldLabel } from 'types/vessel'
 import InfoFieldHistory from './InfoFieldHistory'
+import styles from './Info.module.css'
 
-export enum VesselFieldLabel {
-  name = 'name',
-  flag = 'flag',
-  shipname = 'shipname',
-  firstTransmissionDate = 'firstTransmissionDate',
-  lastTransmissionDate = 'lastTransmissionDate',
-  imo = 'imo',
-  mmsi = 'mmsi',
-  callsign = 'callsign',
-  fleet = 'fleet',
-  origin = 'origin',
-  type = 'type',
-  geartype = 'geartype',
-  length = 'length',
-  depth = 'depth',
-  grossTonnage = 'grossTonnage',
-  owner = 'owner',
-  operator = 'operator',
-  builtYear = 'builtYear',
-  authorizations = 'authorizations',
-  registeredGearType = 'registeredGearType',
-  iuuStatus = 'iuuStatus',
-}
 interface ListItemProps {
   label: VesselFieldLabel
   value?: string
@@ -48,18 +26,16 @@ const InfoField: React.FC<ListItemProps> = ({
   const { t } = useTranslation()
 
   const [modalOpen, setModalOpen] = useState(false)
-  const openModal = useCallback(
-    () => {
-      if (valuesHistory.length > 0) {
-        setModalOpen(true)
-        uaEvent({
-          category: 'Vessel Detail INFO Tab',
-          action: 'Vessel detail INFO tab is open and user click in the history by each field',
-          label: JSON.stringify({[label]: valuesHistory.length}),
-        })
-      } 
-    }, [label, valuesHistory.length]
-  )
+  const openModal = useCallback(() => {
+    if (valuesHistory.length > 0) {
+      setModalOpen(true)
+      uaEvent({
+        category: 'Vessel Detail INFO Tab',
+        action: 'Vessel detail INFO tab is open and user click in the history by each field',
+        label: JSON.stringify({ [label]: valuesHistory.length }),
+      })
+    }
+  }, [label, valuesHistory.length])
   const closeModal = useCallback(() => setModalOpen(false), [])
 
   const since = useMemo(() => valuesHistory.slice(0, 1)?.shift()?.firstSeen, [valuesHistory])