+++ applications/vessel-history/src/features/user/user.slice.ts
@@ -1,5 +1,48 @@
-import { createAsyncThunk } from '@reduxjs/toolkit'
-import GFWAPI from '@globalfishingwatch/api-client'
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
+import { UserData } from '@globalfishingwatch/api-types'
+import GFWAPI, {
+  getAccessTokenFromUrl,
+  removeAccessTokenFromUrl,
+} from '@globalfishingwatch/api-client'
+import { AsyncReducerStatus } from 'utils/async-slice'
+import { RootState } from 'store'
+
+interface UserState {
+  logged: boolean
+  status: AsyncReducerStatus
+  data: UserData | null
+}
+
+const initialState: UserState = {
+  logged: false,
+  status: AsyncReducerStatus.Idle,
+  data: null,
+}
+
+export const GFW_GROUP_ID = 'GFW'
+
+export const fetchUserThunk = createAsyncThunk(
+  'user/fetch',
+  async () => {
+    const accessToken = getAccessTokenFromUrl()
+    if (accessToken) {
+      removeAccessTokenFromUrl()
+    }
+
+    try {
+      return await GFWAPI.login({ accessToken })
+    } catch (e: any) {
+      await logoutUserThunk({ redirectTo: 'gfw-login' })
+      return undefined
+    }
+  },
+  {
+    condition: (_, { getState, extra }) => {
+      const status = selectUserStatus(getState() as RootState)
+      return status !== AsyncReducerStatus.Loading
+    },
+  }
+)
 
 export const logoutUserThunk = createAsyncThunk(
   'user/logout',
@@ -17,3 +60,33 @@ export const logoutUserThunk = createAsyncThunk(
     }
   }
 )
+
+const userSlice = createSlice({
+  name: 'user',
+  initialState,
+  reducers: {},
+  extraReducers: (builder) => {
+    builder.addCase(fetchUserThunk.pending, (state) => {
+      state.status = AsyncReducerStatus.Loading
+    })
+    builder.addCase(fetchUserThunk.fulfilled, (state, action) => {
+      state.status = AsyncReducerStatus.Finished
+      state.logged = true
+      state.data = action.payload ?? null
+    })
+    builder.addCase(fetchUserThunk.rejected, (state) => {
+      state.status = AsyncReducerStatus.Error
+    })
+    builder.addCase(logoutUserThunk.fulfilled, (state) => {
+      state.logged = false
+      state.data = null
+    })
+  },
+})
+
+export const selectUserData = (state: RootState) => state.user.data
+export const selectUserStatus = (state: RootState) => state.user.status
+export const selectUserLogged = (state: RootState) => state.user.logged
+export const isGFWUser = (state: RootState) => state.user.data?.groups.includes(GFW_GROUP_ID)
+
+export default userSlice.reducer