+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-types
 
+## 6.0.1
+
+### Patch Changes
+
+- f5264a92: [WIP] Fishing map/download activity popup
+
+## 6.0.0
+
+### Major Changes
+
+- aa3cf308: Added originalStartDate and originalEndDate to Authorization
+
 ## 5.3.2
 
 ### Patch Changes