+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @globalfishingwatch/react-hooks
 
+## 15.0.1
+
+### Patch Changes
+
+- 1faa75c7: Add geometry on user context layers interaction
+
+## 15.0.0
+
+### Patch Changes
+
+- @globalfishingwatch/api-client@4.0.0
+- @globalfishingwatch/dataviews-client@14.0.0
+- @globalfishingwatch/layer-composer@10.0.0
+
 ## 14.0.1
 
 ### Patch Changes